refactor: Sessions 폴더 이동(shared -> sections)
