+++ pages/2025/index.tsx
@@ -6,7 +6,7 @@ import Overview from '@/components/2025-commingsoon/sections/Overview'
 import Sponsor from '@/components/2025-commingsoon/sections/Sponsor'
 import SNS from '@/components/2025-commingsoon/sections/SNS'
 import Footer from '@/components/2025-commingsoon/sections/Footer'
-import Sessions from '@/components/2025-commingsoon/shared/Sessions'
+import Sessions from '@/components/2025-commingsoon/sections/Sessions'
 import Map from '@/components/2025-commingsoon/sections/Map'
 
 export default function Home() {
